+++ packages/avalanche-module/src/handlers/avalanche-sign-transaction/avalanche-sign-transaction.ts
@@ -16,7 +16,6 @@ import { getProvider } from '../../utils/get-provider';
 import { parseTxDetails } from '../avalanche-send-transaction/utils/parse-tx-details';
 import { getTransactionDetailSections } from '../../utils/get-transaction-detail-sections';
 import { getCoreHeaders } from '@internal/utils';
-import { isDevnet } from '@internal/utils/src/utils/is-devnet';
 
 const GLACIER_API_KEY = process.env.GLACIER_API_KEY;
 
@@ -46,15 +45,14 @@ export const avalancheSignTransaction = async ({
   const vm = Avalanche.getVmByChainAlias(chainAlias);
   const txBytes = utils.hexToBuffer(transactionHex);
   const isTestnet = network.isTestnet ?? false;
-  // TODO(@meeh0w): remove `isDevnet` case after E-upgrade activation on Fuji
-  const provider = await getProvider({ isTestnet, isDevnet: isDevnet(network) });
+  const provider = await getProvider({ isTestnet });
 
   const tx = utils.unpackWithManager(vm, txBytes) as avaxSerial.AvaxTx;
 
   const utxos = await Avalanche.getUtxosByTxFromGlacier({
     transactionHex,
     chainAlias,
-    network: isDevnet(network) ? GlacierNetwork.DEVNET : isTestnet ? GlacierNetwork.FUJI : GlacierNetwork.MAINNET,
+    network: isTestnet ? GlacierNetwork.FUJI : GlacierNetwork.MAINNET,
     url: glacierApiUrl,
     token: GLACIER_API_KEY,
     headers: getCoreHeaders(appInfo),