+++ packages/avalanche-module/src/handlers/get-network-fee/get-network-fee.ts
@@ -6,14 +6,12 @@ import { getProvider } from '../../utils/get-provider';
  */
 export async function getNetworkFee({
   isTestnet,
-  isDevnet,
   vmName,
 }: {
   isTestnet: boolean;
-  isDevnet?: boolean; // TODO(@meeh0w): remove `isDevnet` after E-upgrade activation on Fuji
   vmName: NetworkVMType;
 }): Promise<NetworkFees> {
-  const provider = await getProvider({ isTestnet, isDevnet });
+  const provider = await getProvider({ isTestnet });
 
   // Return static fees for X-Chain and pre-Etna P-Chain
   if (vmName === NetworkVMType.AVM || !provider.isEtnaEnabled()) {