+++ packages/avalanche-module/src/module.ts
@@ -13,7 +13,7 @@ import type {
   ConstructorParams,
   AppInfo,
 } from '@avalabs/vm-module-types';
-import { Environment, parseManifest, RpcMethod } from '@avalabs/vm-module-types';
+import { parseManifest, RpcMethod } from '@avalabs/vm-module-types';
 import { rpcErrors } from '@metamask/rpc-errors';
 import ManifestJson from '../manifest.json';
 import { getNetworkFee } from './handlers/get-network-fee/get-network-fee';
@@ -29,7 +29,6 @@ import { avalancheSendTransaction } from './handlers/avalanche-send-transaction/
 import { avalancheSignTransaction } from './handlers/avalanche-sign-transaction/avalanche-sign-transaction';
 import type { Avalanche } from '@avalabs/core-wallets-sdk';
 import { getProvider } from './utils/get-provider';
-import { isDevnet } from '@internal/utils/src/utils/is-devnet';
 
 export class AvalancheModule implements Module {
   #glacierService: AvalancheGlacierService;
@@ -51,7 +50,7 @@ export class AvalancheModule implements Module {
   }
 
   getProvider(network: Network): Promise<Avalanche.JsonRpcProvider> {
-    return getProvider({ isTestnet: Boolean(network.isTestnet), isDevnet: isDevnet(network) });
+    return getProvider({ isTestnet: Boolean(network.isTestnet) });
   }
 
   getAddress({ accountIndex, xpubXP, walletType, network }: GetAddressParams): Promise<GetAddressResponse> {
@@ -64,9 +63,7 @@ export class AvalancheModule implements Module {
   }
 
   getBalances({ addresses, network, storage, currency }: GetBalancesParams): Promise<GetBalancesResponse> {
-    // TODO(@meeh0w): remove `isDevnet` case after E-upgrade activation on Fuji
-    const proxyApiUrl = isDevnet(network) ? getEnv(Environment.DEV).proxyApiUrl : this.#proxyApiUrl;
-    const tokenService = new TokenService({ storage, proxyApiUrl });
+    const tokenService = new TokenService({ storage, proxyApiUrl: this.#proxyApiUrl });
     return getBalances({ addresses, currency, network, glacierService: this.#glacierService, tokenService });
   }
 
@@ -77,7 +74,7 @@ export class AvalancheModule implements Module {
 
   getNetworkFee(network: Network): Promise<NetworkFees> {
     const { isTestnet, vmName } = network;
-    return getNetworkFee({ isTestnet: Boolean(isTestnet), isDevnet: isDevnet(network), vmName });
+    return getNetworkFee({ isTestnet: Boolean(isTestnet), vmName });
   }
 
   getTransactionHistory({ network, address, nextPageToken, offset }: GetTransactionHistory) {